requirements.txt
flask==2.2.2
flask-wtf==1.0.1
WTForms~=3.0.1
email-validator==1.3.0
flask_sqlalchemy==3.0.2
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-mail==0.9.1
Pillow==9.3.0
pylint==2.15.5
itsdangerous==2.0.1